feat: moved erc20 recover to a dedicated contract
